Refactor test_no_renewal_with_hooks() - Fix #45
